save worlds, open worlds, send tiles

+++ src/raw_id.rs
@@ -2,4 +2,4 @@ use bevy::prelude::*;
 use serde::{Deserialize, Serialize};
 
 #[derive(Debug, Copy, Clone, Serialize, Deserialize, PartialEq, Eq, Hash, Component)]
-pub struct RawId(pub u64);
+pub struct RawId(pub u32);
